* Added support for non-rounded ages

+++ src/models/VersionData/schema.js
@@ -2,7 +2,7 @@ export const schema = `
     type VersionData {
         version: String!,
         date: String,
-        age_days: Int!,
+        age_days: Float!,
         age_text: String!,
         description: String!,
     }
